+++ react/__tests__/Arrows.test.tsx
@@ -4,6 +4,7 @@ import { render, fireEvent } from '@vtex/test-tools/react'
 import { useSliderState } from '../components/SliderContext'
 import Arrow, { CSS_HANDLES } from '../components/Arrow'
 import { useContextCssHandles } from '../modules/cssHandles'
+import { mockedUseContextCssHandlesFn } from '../__fixtures__/CssHandlesHelper'
 
 const mockGoForward = jest.fn()
 const mockGoBack = jest.fn()
@@ -29,14 +30,9 @@ const mockedUseContextCssHandles = useContextCssHandles as jest.Mock<
   ReturnType<typeof useContextCssHandles>
 >
 
-mockedUseContextCssHandles.mockImplementation(() => ({
-  handles: CSS_HANDLES.reduce<Record<string, string>>((acc, handle) => {
-    acc[handle] = handle
-
-    return acc
-  }, {}),
-  withModifiers: (handle, modifier) => `${handle} ${handle}--${modifier}`,
-}))
+mockedUseContextCssHandles.mockImplementation(() =>
+  mockedUseContextCssHandlesFn(CSS_HANDLES)
+)
 
 beforeEach(() => {
   mockGoForward.mockClear()