+++ react/__tests__/PaginationDots.test.tsx
@@ -3,6 +3,7 @@ import { render, fireEvent } from '@vtex/test-tools/react'
 
 import PaginationDots, { CSS_HANDLES } from '../components/PaginationDots'
 import { useContextCssHandles } from '../modules/cssHandles'
+import { mockedUseContextCssHandlesFn } from '../__fixtures__/CssHandlesHelper'
 
 const mockGoForward = jest.fn()
 const mockGoBack = jest.fn()
@@ -50,14 +51,9 @@ const mockedUseContextCssHandles = useContextCssHandles as jest.Mock<
   ReturnType<typeof useContextCssHandles>
 >
 
-mockedUseContextCssHandles.mockImplementation(() => ({
-  handles: CSS_HANDLES.reduce<Record<string, string>>((acc, handle) => {
-    acc[handle] = handle
-
-    return acc
-  }, {}),
-  withModifiers: (handle, modifier) => `${handle} ${handle}--${modifier}`,
-}))
+mockedUseContextCssHandles.mockImplementation(() =>
+  mockedUseContextCssHandlesFn(CSS_HANDLES)
+)
 
 describe('Basic rendering', () => {
   it('should render the correct number of pagination dots based on the number of slider pages', () => {