+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Prop `classes`to SliderLayout.
+- Prop `classes` to SliderLayout.
 
 ### Changed
 - Migrate to `vtex.css-handles@1.x`.