+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - New `centerModeSlidesGap` prop.
+
 ## [0.19.1] - 2021-06-17
 
 ### Fixed